+++ pkg/index/index.go
@@ -54,6 +54,7 @@ type Field struct {
 	Term   []byte
 	Key    FieldKey
 	NoSort bool
+	Store  bool
 }
 
 // RangeOpts contains options to performance a continuous scan.
@@ -91,17 +92,18 @@ func (r RangeOpts) Between(value []byte) int {
 	return 0
 }
 
-// ItemRef represents a reference to an item.
-type ItemRef struct {
-	Term      []byte
-	SeriesID  common.SeriesID
-	DocID     uint64
-	Timestamp int64
+// DocumentResult represents a document in a index.
+type DocumentResult struct {
+	Values      map[string][]byte
+	SortedValue []byte
+	SeriesID    common.SeriesID
+	DocID       uint64
+	Timestamp   int64
 }
 
 // SortedField returns the value of the sorted field.
-func (ir ItemRef) SortedField() []byte {
-	return ir.Term
+func (ir DocumentResult) SortedField() []byte {
+	return ir.SortedValue
 }
 
 // FieldIterator allows iterating over a field's posting values.
@@ -120,8 +122,8 @@ func (i *dummyIterator) Next() bool {
 	return false
 }
 
-func (i *dummyIterator) Val() *ItemRef {
-	return &ItemRef{}
+func (i *dummyIterator) Val() *DocumentResult {
+	return &DocumentResult{}
 }
 
 func (i *dummyIterator) Close() error {
@@ -151,8 +153,8 @@ type Writer interface {
 
 // FieldIterable allows building a FieldIterator.
 type FieldIterable interface {
-	Iterator(fieldKey FieldKey, termRange RangeOpts, order modelv1.Sort, preLoadSize int) (iter FieldIterator[*ItemRef], err error)
-	Sort(sids []common.SeriesID, fieldKey FieldKey, order modelv1.Sort, timeRange *timestamp.TimeRange, preLoadSize int) (FieldIterator[*ItemRef], error)
+	Iterator(fieldKey FieldKey, termRange RangeOpts, order modelv1.Sort, preLoadSize int) (iter FieldIterator[*DocumentResult], err error)
+	Sort(sids []common.SeriesID, fieldKey FieldKey, order modelv1.Sort, timeRange *timestamp.TimeRange, preLoadSize int) (FieldIterator[*DocumentResult], error)
 }
 
 // Searcher allows searching a field either by its key or by its key and term.
@@ -180,6 +182,12 @@ type Series struct {
 
 func (s Series) String() string {
 	return fmt.Sprintf("%s:%d", s.EntityValues, s.ID)
+}
+
+// SeriesDocument represents a series document in a index.
+type SeriesDocument struct {
+	Fields map[string][]byte
+	Key    Series
 }
 
 // SeriesMatcherType represents the type of series matcher.