+++ pkg/index/inverted/sort.go
@@ -36,7 +36,7 @@ import (
 // Sort creates a sorted iterator based on a given set of seriesIDs and field key, with specified sorting order, time range and preload size.
 func (s *Store) Sort(sids []common.SeriesID, fieldKey index.FieldKey, order modelv1.Sort,
 	timeRange *timestamp.TimeRange, preLoadSize int,
-) (iter index.FieldIterator[*index.ItemRef], err error) {
+) (iter index.FieldIterator[*index.DocumentResult], err error) {
 	reader, err := s.writer.Reader()
 	if err != nil {
 		return nil, err
@@ -70,27 +70,23 @@ func (s *Store) Sort(sids []common.SeriesID, fieldKey index.FieldKey, order mode
 		sortedKey = "-" + sortedKey
 	}
 	result := &sortIterator{
-		query:       query,
-		reader:      reader,
-		sortedKey:   sortedKey,
-		sortedField: fk,
-		size:        preLoadSize,
-		sid:         fieldKey.SeriesID,
+		query:     query,
+		reader:    reader,
+		sortedKey: sortedKey,
+		size:      preLoadSize,
 	}
 	return result, nil
 }
 
 type sortIterator struct {
-	query       bluge.Query
-	err         error
-	reader      *bluge.Reader
-	current     *blugeMatchIterator
-	closer      *run.Closer
-	sortedKey   string
-	sortedField string
-	size        int
-	skipped     int
-	sid         common.SeriesID
+	query     bluge.Query
+	err       error
+	reader    *bluge.Reader
+	current   *blugeMatchIterator
+	closer    *run.Closer
+	sortedKey string
+	size      int
+	skipped   int
 }
 
 func (si *sortIterator) Next() bool {
@@ -125,7 +121,7 @@ func (si *sortIterator) loadCurrent() bool {
 		return false
 	}
 
-	iter := newBlugeMatchIterator(documentMatchIterator, nil, []string{si.sortedField})
+	iter := newBlugeMatchIterator(documentMatchIterator, nil, nil)
 	si.current = &iter
 	if si.next() {
 		return true
@@ -142,18 +138,12 @@ func (si *sortIterator) next() bool {
 	return false
 }
 
-func (si *sortIterator) Val() *index.ItemRef {
+func (si *sortIterator) Val() *index.DocumentResult {
 	v := si.current.Val()
-	sv, ok := v.values[si.sortedField]
-	if !ok {
+	if v.SortedValue == nil {
 		panic("sorted field not found in document")
 	}
-	return &index.ItemRef{
-		SeriesID:  v.seriesID,
-		DocID:     v.docID,
-		Term:      sv,
-		Timestamp: v.timestamp,
-	}
+	return &v
 }
 
 func (si *sortIterator) Close() error {